Removed all except 1d array
