+++ src/libplctag/TagOfT.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace libplctag
 {
-    public class Tag2d<Marshaller, T>
+    public class Tag<Marshaller, T>
         where Marshaller : IMarshaller<T>, new()
     {
 
@@ -21,19 +21,17 @@ namespace libplctag
         /// <param name="path">Required for LGX, Optional for PLC/SLC/MLGX IOI path to access the PLC from the gateway.
         /// <param name="plcType">Allen-Bradley CPU model</param>
         /// <param name="name">The textual name of the tag to access. The name is anything allowed by the protocol. E.g. myDataStruct.rotationTimer.ACC, myDINTArray[42] etc.</param>
-        /// <param name="dimension1Length">Length of array dimension 1</param>
-        /// <param name="dimension2Length">Length of array dimension 2</param>
+        /// <param name="elementCount">elements count: 1- single, n-array.</param>
         /// <param name="millisecondTimeout"></param>
         /// <param name="protocol">Currently only ab_eip supported.</param>
         /// <param name="readCacheMillisecondDuration">Set the amount of time to cache read results</param>
         /// <param name="useConnectedMessaging">Control whether to use connected or unconnected messaging.</param>
-        public Tag2d(IPAddress gateway,
+        public Tag(IPAddress gateway,
                    string path,
                    PlcType plcType,
                    string name,
-                   int dimension1Length,
-                   int dimension2Length,
                    int millisecondTimeout,
+                   int elementCount,
                    Protocol protocol = Protocol.ab_eip,
                    int readCacheMillisecondDuration = default,
                    bool useConnectedMessaging = true)
@@ -44,9 +42,6 @@ namespace libplctag
                 PlcType = plcType
             };
 
-            Dimension1Length = dimension1Length;
-            Dimension2Length = dimension2Length;
-
             _tag = new Tag(
                 gateway,
                 path,
@@ -54,12 +49,13 @@ namespace libplctag
                 _marshaller.ElementSize,
                 name,
                 millisecondTimeout,
-                dimension1Length * dimension2Length,
+                elementCount,
                 protocol,
                 readCacheMillisecondDuration,
                 useConnectedMessaging);
 
-            Value = new T[Dimension1Length, Dimension2Length];
+
+            Value = new T[elementCount];
             DecodeAll();
         }
 
@@ -67,8 +63,7 @@ namespace libplctag
         public IPAddress Gateway => _tag.Gateway;
         public string Path => _tag.Path;
         public PlcType PlcType => _tag.PlcType;
-        public int Dimension1Length { get; }
-        public int Dimension2Length { get; }
+        public int Count => _tag.ElementCount;
         public string Name => _tag.Name;
         public bool UseConnectedMessaging => _tag.UseConnectedMessaging;
         public int ReadCacheMillisecondDuration
@@ -76,7 +71,6 @@ namespace libplctag
             get => _tag.ReadCacheMillisecondDuration;
             set => _tag.ReadCacheMillisecondDuration = value;
         }
-
 
         public async Task ReadAsync(int millisecondTimeout, CancellationToken token = default)
         {
@@ -114,25 +108,23 @@ namespace libplctag
             _tag.Write(millisecondTimeout);
         }
 
-        int GetOffset(int i, int j) => _marshaller.ElementSize * (i * Dimension2Length + j);
+        int GetOffset(int i) => _marshaller.ElementSize * i;
 
         void DecodeAll()
         {
-            for (int ii = 0; ii < Dimension1Length; ii++)
-                for (int jj = 0; jj < Dimension2Length; jj++)
-                    Value[ii, jj] = _marshaller.Decode(_tag,  GetOffset(ii, jj));
+            for (int ii = 0; ii < Value.Length; ii++)
+                Value[ii] = _marshaller.Decode(_tag, GetOffset(ii));
         }
 
         void EncodeAll()
         {
-            for (int ii = 0; ii < Dimension1Length; ii++)
-                for (int jj = 0; jj < Dimension2Length; jj++)
-                    _marshaller.Encode(_tag, GetOffset(ii, jj), Value[ii, jj]);
+            for (int ii = 0; ii < Value.Length; ii++)
+                _marshaller.Encode(_tag, GetOffset(ii), Value[ii]);
         }
 
         public Status GetStatus() => _tag.GetStatus();
 
-        public T[,] Value { get; set; }
+        public T[] Value { get; set; }
 
     }
 }
